Add cache threshold for populate
Avoids unnecessary running of Populate script
